Develop

+++ src/main/node/process.ts
@@ -1,4 +1,4 @@
-/* eslint-disable quotes */
+/* eslint-disable quotes,no-await-in-loop */
 import fse from 'fs-extra'
 import path from "path"
 import {
@@ -9,8 +9,9 @@ import {PhrasesStatCollector} from '../common/PhrasesStatCollector'
 import {WordsCache} from '../common/WordsCache'
 import {WordsStat} from '../common/WordsStat'
 import {txtBookBufferToString, xmlBookBufferToString} from './helpers'
-import {processFiles} from './processFiles'
+import {processArchiveTarXz, processFiles} from './processFiles'
 import readline from 'readline'
+import papaparse from 'papaparse'
 
 export const wordsPerPage = 200
 
@@ -45,7 +46,7 @@ export async function calcStat({
 
 	await processFiles({
 		fileOrDirPath,
-		alwaysReadBuffer: true,
+		readBuffer: true,
 		processArchives : true,
 		filterPaths(isDir, archivePath, _fileOrDirPath) {
 			if (!isDir && !/\.(txt|fb2)$/i.test(_fileOrDirPath)) {
@@ -53,7 +54,7 @@ export async function calcStat({
 			}
 			return true
 		},
-		async processFile(rootDir, archivePath, filePath, buffer) {
+		async processFile(rootDir, archivePath, filePath, stream, buffer) {
 			const text = xmlBookBufferToString(buffer)
 			const totalWords = phrasesStatCollector.addText(text)
 			if (onFileHandled) {
@@ -71,26 +72,28 @@ export async function calcStat({
 export async function calcWordStat({
 	wordRegExp,
 	fileOrDirPath,
+	filterPaths,
 	bufferToString,
 }: {
 	wordRegExp: RegExp,
 	fileOrDirPath: string,
+	filterPaths?: (isDir: boolean, archivePath: string, fileOrDirPath: string) => boolean,
 	bufferToString: (buffer: Buffer) => Promise<string>|string,
 }) {
 	const wordsStat = new WordsStat()
 	await processFiles({
 		fileOrDirPath,
-		async processFile(filePath) {
-			if (!/\.(txt|fb2)$/i.test(filePath)) {
-				return
-			}
-			const buffer = await fse.readFile(filePath)
+		readBuffer: true,
+		processArchives : true,
+		filterPaths,
+		async processFile(rootDir, archivePath, filePath, stream, buffer) {
 			const text = await bufferToString(buffer)
 
 			// let totalWords = 0
 			parsePhrases(text, wordRegExp, word => {
-				wordsStat.add(word.toLowerCase())
 				// totalWords++
+				word = word.toLowerCase()
+				wordsStat.add(word)
 			})
 
 			// console.log('totalWords:', totalWords)
@@ -100,11 +103,16 @@ export async function calcWordStat({
 	return wordsStat
 }
 
-interface IAnalyzeBookResult {
+export interface IBookStat {
 	unknownWordsIn3Pages: number,
 	unknownWordsIn20Pages: number,
 	unknownWords: number,
 	totalWords: number,
+}
+
+export interface ILibgenBookStat extends IBookStat {
+	id: number,
+	hash: string,
 }
 
 export function analyzeBook({
@@ -115,15 +123,16 @@ export function analyzeBook({
 	text: string,
 	wordRegExp: RegExp,
 	wordFilter: (word: string) => boolean,
-}): IAnalyzeBookResult {
+}): IBookStat {
 	const wordStat = new WordsStat()
 	let totalWords = 0
 	parsePhrases(text, wordRegExp, word => {
+		totalWords++
+		word = word.toLowerCase()
 		if (wordFilter && !wordFilter(word)) {
 			return
 		}
 		wordStat.add(word)
-		totalWords++
 	})
 
 	if (!totalWords) {
@@ -152,64 +161,12 @@ export function analyzeBook({
 	}
 }
 
-export async function analyzeBooks<TLogEntry extends IAnalyzeBookResult>({
-	booksDir,
-	resultsDir,
-	totalBooks,
-	reportHeader,
-	logEntryToReportLine,
-	filterPaths,
-	analyzeBook: _analyzeBook,
-}: {
-	booksDir: string,
-	resultsDir: string,
-	totalBooks?: number,
-	reportHeader: string,
-	logEntryToReportLine: (logEntry: TLogEntry) => string,
-	filterPaths?: (isDir: boolean, archivePath: string, fileOrDirPath: string) => boolean,
-	analyzeBook: (
-		rootDir: string,
-		archivePath: string,
-		filePath: string,
-		buffer?: Buffer,
-	) => Promise<TLogEntry>|TLogEntry,
-}) {
-	const stateFile = path.resolve(resultsDir, 'state.json')
-	const logFile = path.resolve(resultsDir, 'log.json')
-	const reportFile = path.resolve(resultsDir, 'report.txt')
-	const dir = path.dirname(reportFile)
-
-	if (!fse.existsSync(dir)) {
-		await fse.mkdirp(dir)
-	}
-
-	const state: {
-		processedFiles: { [key: string]: number }
-	} = fse.existsSync(stateFile)
-		? await fse.readJSON(stateFile, { encoding: 'utf-8' })
-		: { processedFiles: {} }
-
-	const log: TLogEntry[] = []
-
-	let processedBooks = Object.keys(state.processedFiles).reduce((a, o) => {
-		return a + state.processedFiles[o]
-	}, 0)
-
-	function showProgress() {
-		if (totalBooks) {
-			console.log(`${processedBooks} / (${(processedBooks * 100 / totalBooks).toFixed(2)}%)`)
-		} else {
-			console.log(processedBooks)
-		}
-	}
-
-	async function report() {
-		if (!fse.existsSync(logFile)) {
-			return
-		}
-		const logStr = (await fse.readFile(logFile, {encoding: 'utf-8'})) + ']'
-		const _log: TLogEntry[] = JSON.parse(logStr)
-		_log.sort((o1, o2) => {
+export async function readBookStats<TLogEntry extends IBookStat>(bookStatsFile: string) {
+	const bookStatsStr = (await fse.readFile(bookStatsFile, {encoding: 'utf-8'})) + ']'
+	let bookStats: TLogEntry[] = JSON.parse(bookStatsStr)
+	bookStats = bookStats
+		.filter(o => o.totalWords / wordsPerPage >= 2)
+		.sort((o1, o2) => {
 			if (o1.unknownWordsIn3Pages !== o2.unknownWordsIn3Pages) {
 				return o1.unknownWordsIn3Pages > o2.unknownWordsIn3Pages ? 1 : -1
 			}
@@ -224,21 +181,79 @@ export async function analyzeBooks<TLogEntry extends IAnalyzeBookResult>({
 			}
 			return 0
 		})
+	return bookStats
+}
 
+export async function analyzeBooks<TBookStat extends IBookStat>({
+	booksDir,
+	resultsDir,
+	totalBooks,
+	reportHeader,
+	bookStatToReportLine,
+	filterPaths,
+	analyzeBook: _analyzeBook,
+}: {
+	booksDir: string,
+	resultsDir: string,
+	totalBooks?: number,
+	reportHeader: string,
+	bookStatToReportLine: (bookStat: TBookStat) => string,
+	filterPaths?: (isDir: boolean, archivePath: string, fileOrDirPath: string) => boolean,
+	analyzeBook: (
+		rootDir: string,
+		archivePath: string,
+		filePath: string,
+		buffer?: Buffer,
+	) => Promise<TBookStat>|TBookStat,
+}) {
+	const stateFile = path.resolve(resultsDir, 'state.json')
+	const bookStatsFile = path.resolve(resultsDir, 'stat.json')
+	const reportFile = path.resolve(resultsDir, 'report.txt')
+	const dir = path.dirname(reportFile)
+
+	if (!fse.existsSync(dir)) {
+		await fse.mkdirp(dir)
+	}
+
+	const state: {
+		processedFiles: { [key: string]: number }
+	} = fse.existsSync(stateFile)
+		? await fse.readJSON(stateFile, { encoding: 'utf-8' })
+		: { processedFiles: {} }
+
+	const bookStats: TBookStat[] = []
+
+	let processedBooks = Object.keys(state.processedFiles).reduce((a, o) => {
+		return a + state.processedFiles[o]
+	}, 0)
+
+	function showProgress() {
+		if (totalBooks) {
+			console.log(`${processedBooks} / (${(processedBooks * 100 / totalBooks).toFixed(2)}%)`)
+		} else {
+			console.log(processedBooks)
+		}
+	}
+
+	async function report() {
+		if (!fse.existsSync(bookStatsFile)) {
+			return
+		}
+		const _bookStats = await readBookStats<TBookStat>(bookStatsFile)
 		let reportStr = reportHeader + '\r\n'
-		reportStr += _log.map(logEntryToReportLine)
+		reportStr += _bookStats.map(bookStatToReportLine)
 			.join('\r\n')
 
 		await fse.writeFile(reportFile, reportStr, { encoding: 'utf-8' })
 	}
 
 	async function save() {
-		if (log.length > 0) {
-			let logStr = JSON.stringify(log)
-			logStr = (fse.existsSync(logFile) ? ',' : '[')
-				+ logStr.substring(1, logStr.length - 1)
-			await fse.appendFile(logFile, logStr, {encoding: 'utf-8'})
-			log.length = 0
+		if (bookStats.length > 0) {
+			let bookStatsStr = JSON.stringify(bookStats)
+			bookStatsStr = (fse.existsSync(bookStatsFile) ? ',' : '[')
+				+ bookStatsStr.substring(1, bookStatsStr.length - 1)
+			await fse.appendFile(bookStatsFile, bookStatsStr, {encoding: 'utf-8'})
+			bookStats.length = 0
 		}
 
 		await fse.writeJSON(stateFile, state, { encoding: 'utf-8' })
@@ -250,27 +265,33 @@ export async function analyzeBooks<TLogEntry extends IAnalyzeBookResult>({
 
 	showProgress()
 
+	let prevTime = Date.now()
+
 	await processFiles({
 		fileOrDirPath   : booksDir,
 		processArchives : true,
-		alwaysReadBuffer: true,
+		readBuffer      : true,
 		filterPaths(isDir, archivePath, fileOrDirPath) {
 			if (!isDir && !archivePath && fileOrDirPath in state.processedFiles) {
 				return false
 			}
-			return filterPaths(isDir, archivePath, fileOrDirPath)
+			return !filterPaths || filterPaths(isDir, archivePath, fileOrDirPath)
 		},
-		async processFile(rootDir, archivePath, filePath, buffer) {
-			const logEntry = await _analyzeBook(rootDir, archivePath, filePath, buffer)
-			if (logEntry) {
-				log.push(logEntry)
+		async processFile(rootDir, archivePath, filePath, stream, buffer) {
+			const bookStat = await _analyzeBook(rootDir, archivePath, filePath, buffer)
+			if (bookStat) {
+				bookStats.push(bookStat)
 			}
 		},
 		async onFileProcessed(rootDir, archivePath, filePath) {
 			if (!archivePath) {
-				state.processedFiles[filePath] = log.length
-				processedBooks += log.length
-				await save()
+				const now = Date.now()
+				if (now > prevTime + 60 * 1000) {
+					prevTime = now
+					state.processedFiles[filePath] = bookStats.length
+					processedBooks += bookStats.length
+					await save()
+				}
 			}
 		},
 	})
@@ -278,23 +299,7 @@ export async function analyzeBooks<TLogEntry extends IAnalyzeBookResult>({
 	await save()
 }
 
-export async function processLibgen({
-	dbPath,
-	booksDir,
-	resultsDir,
-	wordRegExp,
-	wordFilter,
-	filterPaths,
-}: {
-	dbPath: string,
-	booksDir: string,
-	resultsDir: string,
-	wordRegExp: RegExp,
-	wordFilter: (word: string) => boolean,
-	filterPaths?: (isDir: boolean, archivePath: string, fileOrDirPath: string) => boolean,
-}) {
-	// region parse db
-
+export async function parseLibgenDb(dbPath: string) {
 	let books: {
 		[hash: string]: {
 			id: number
@@ -325,7 +330,11 @@ export async function processLibgen({
 				continue
 			}
 			if (line) {
-				let [idStr, hash, lang, , author, title, seriesStr] = line.split(',')
+				let [idStr, hash, lang, , author, title, seriesStr] = papaparse.parse(line, {
+					header: false,
+				}).data[0] as string[]
+				// let [idStr, hash, lang, , author, title, seriesStr] = line.split(',')
+
 				idStr = idStr.trim()
 				seriesStr = seriesStr.trim()
 				lang = lang.toLowerCase().trim()
@@ -345,31 +354,46 @@ export async function processLibgen({
 		dbStream.close()
 	}
 
-	// endregion
+	return books
+}
+
+export async function processLibgen({
+	dbPath,
+	booksDir,
+	resultsDir,
+	wordRegExp,
+	wordFilter,
+	filterPaths,
+}: {
+	dbPath: string,
+	booksDir: string,
+	resultsDir: string,
+	wordRegExp: RegExp,
+	wordFilter: (word: string) => boolean,
+	filterPaths?: (isDir: boolean, archivePath: string, fileOrDirPath: string) => boolean,
+}) {
+	const books = await parseLibgenDb(dbPath)
 
 	const totalBooks = Object.keys(books).length
 
-	await analyzeBooks<{
-		id: number,
-		hash: string,
-	} & IAnalyzeBookResult>({
+	await analyzeBooks<ILibgenBookStat>({
 		booksDir,
 		resultsDir,
 		totalBooks,
 		reportHeader: 'id\thash\tunknownWordsIn3Pages\tunknownWordsIn20Pages\tunknownWords\ttotalPages',
-		logEntryToReportLine(logEntry) {
+		bookStatToReportLine(bookStat) {
 			return `${
-				logEntry.id
+				bookStat.id
 			}\t${
-				logEntry.hash
+				bookStat.hash
 			}\t${
-				logEntry.unknownWordsIn3Pages
+				bookStat.unknownWordsIn3Pages
 			}\t${
-				logEntry.unknownWordsIn20Pages
+				bookStat.unknownWordsIn20Pages
 			}\t${
-				logEntry.unknownWords
+				bookStat.unknownWords
 			}\t${
-				logEntry.totalWords / wordsPerPage
+				bookStat.totalWords / wordsPerPage
 			}`
 		},
 		filterPaths,
@@ -382,16 +406,16 @@ export async function processLibgen({
 
 			const text = txtBookBufferToString(buffer)
 
-			const logEntry = analyzeBook({
+			const bookStat = analyzeBook({
 				text,
 				wordRegExp,
 				wordFilter,
 			})
 
-			return {
+			return bookStat && {
+				...bookStat,
 				id: book.id,
 				hash,
-				...logEntry,
 			}
 		},
 	})
@@ -410,32 +434,107 @@ export async function processBooks({
 	wordFilter: (word: string) => boolean,
 	filterPaths?: (isDir: boolean, archivePath: string, fileOrDirPath: string) => boolean,
 }) {
-	await analyzeBooks({
+	await analyzeBooks<{
+		filePath,
+	} & IBookStat>({
 		booksDir,
 		resultsDir,
 		reportHeader: 'unknownWordsIn3Pages\tunknownWordsIn20Pages\tunknownWords\ttotalPages',
-		logEntryToReportLine(logEntry) {
+		bookStatToReportLine(bookStat) {
 			return `${
-				logEntry.unknownWordsIn3Pages
+				bookStat.unknownWordsIn3Pages
 			}\t${
-				logEntry.unknownWordsIn20Pages
+				bookStat.unknownWordsIn20Pages
 			}\t${
-				logEntry.unknownWords
+				bookStat.unknownWords
 			}\t${
-				logEntry.totalWords / wordsPerPage
+				bookStat.totalWords / wordsPerPage
+			}\t${
+				bookStat.filePath
 			}`
 		},
 		filterPaths,
 		analyzeBook(rootDir, archivePath, filePath, buffer) {
 			const text = xmlBookBufferToString(buffer)
 
-			const logEntry = analyzeBook({
+			const bookStat = analyzeBook({
 				text,
 				wordRegExp,
 				wordFilter,
 			})
 
-			return logEntry
+			return bookStat && {
+				...bookStat,
+				filePath,
+			}
 		},
 	})
+}
+
+export async function libgenUnpack({
+	dbPath,
+	booksDir,
+	bookStats,
+	unpackDir,
+}: {
+	dbPath: string,
+	booksDir: string,
+	bookStats: ILibgenBookStat[],
+	unpackDir: string,
+}) {
+	const archives: {
+		[archiveName: string]: {
+			[hash: string]: ILibgenBookStat
+		}
+	} = bookStats
+		.reduce((a, o) => {
+			const archiveName = Math.floor(o.id / 1000) + '000.tar.xz'
+			let stats = a[archiveName]
+			if (!stats) {
+				a[archiveName] = stats = {}
+			}
+			stats[o.hash] = o
+			return a
+		}, {})
+
+	const books = await parseLibgenDb(dbPath)
+
+	if (!fse.existsSync(unpackDir)) {
+		await fse.mkdirp(unpackDir)
+	}
+
+	for (const archiveName in archives) {
+		if (Object.prototype.hasOwnProperty.call(archives, archiveName)) {
+			const stats = archives[archiveName]
+			await processArchiveTarXz({
+				archivePath: path.resolve(booksDir, archiveName),
+				async processFile(archivePath, innerFilePath, stream) {
+					const hash = innerFilePath.match(/\/(\w+)(?:\.\w+)?$/)?.[1]?.toLowerCase()
+					const bookStat = stats[hash]
+					if (!bookStat) {
+						return
+					}
+
+					const book = books[bookStat.hash]
+
+					let fileName = book
+						? (book.author || '')
+							+ ' - ' + (book.title || '')
+							+ ' ' + (book.series || '')
+						: bookStat.hash
+
+					fileName = fileName
+						.replace(/[\\/:*?<>|"]/g, '')
+						.replace(/\s+/g, ' ')
+						.trim()
+
+					fileName = Math.round(bookStat.unknownWordsIn3Pages * 100).toString().padStart(5, '0') + ' - ' + fileName
+
+					const filePath = path.resolve(unpackDir, fileName + path.extname(innerFilePath))
+
+					await stream.pipe(fse.createWriteStream(filePath))
+				},
+			})
+		}
+	}
 }
